Split IBaseEntityService into its own file
